Move TZ to top level
