+++ core/__init__.py
@@ -8,4 +8,3 @@ exchange.
 from core.market import Market
 from core.MarketAPI import MarketAPI
 from core.markets import SimulatedMarket, GeminiMarket
-from core.misc import TZ_NAME, TZ
