+++ misc.py
