+++ primitives/cache.py
@@ -1,8 +1,8 @@
-import logging
-
-from core.misc import TZ
 from datetime import datetime, timedelta
+import logging
 from typing import Any, Callable, Tuple, Dict
+
+from misc import TZ
 
 
 class CachedValue(object):
